requirements.txt
fastapi==0.111.1
uvicorn
jinja2
python-dotenv
httpx
ollama
requests
pydantic<2
aiohttp
python-multipart
google-generativeai
openai
langchain
translators
